requirements.txt
python-dotenv==1.0.0
playsound==1.3.0
openai==1.4.0
pygobject==3.46.0